Split into two modules, based on dependencies
